Migrate activityHome to fragment Home and create view for profile
