+++ app/src/main/java/com/peakDevCol/gympartner/domain/model/ProfileModel.kt
@@ -1,0 +1,3 @@
+package com.peakDevCol.gympartner.domain.model
+
+data class ProfileModel(val profileData: List<String>)
