+++ app/src/main/java/com/peakDevCol/gympartner/ui/profile/ProfileFragment.kt
@@ -1,0 +1,25 @@
+package com.peakDevCol.gympartner.ui.profile
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.peakDevCol.gympartner.R
+import com.peakDevCol.gympartner.databinding.FragmentProfileBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+class ProfileFragment : Fragment() {
+
+    private lateinit var binding: FragmentProfileBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState:Bundle?
+    ): View {
+        binding = FragmentProfileBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+}
